delete examples
