--- 351004/Kuchko/Distcomp/Distcomp.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="Distcomp/Distcomp.csproj" />
-</Solution>
