--- 351004/Kuchko/Distcomp/Distcomp/Distcomp.http
@@ -1,4 +1,0 @@
-@Distcomp_HostAddress = http://localhost:5028
-
-GET {{Distcomp_HostAddress}}/weatherforecast/
-Accept: application/json
