+++ 351004/Kuchko/ServerApp/.gitattributes
@@ -1,0 +1,2 @@
+*.sh text eol=lf
+*.cmd text eol=crlf
