+++ 351004/Kuchko/ServerApp/ServerApp.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerApp", "ServerApp\ServerApp.csproj", "{F2EFEDB0-911A-4C64-8CE1-CEC4D2887887}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F2EFEDB0-911A-4C64-8CE1-CEC4D2887887}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F2EFEDB0-911A-4C64-8CE1-CEC4D2887887}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F2EFEDB0-911A-4C64-8CE1-CEC4D2887887}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F2EFEDB0-911A-4C64-8CE1-CEC4D2887887}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
