+++ 351004/Kuchko/ServerApp/ServerApp/Dockerfile
@@ -1,0 +1,23 @@
+﻿FROM mcr.microsoft.com/dotnet/aspnet:9.0 AS base
+USER $APP_UID
+WORKDIR /app
+EXPOSE 8080
+EXPOSE 8081
+
+FROM mcr.microsoft.com/dotnet/sdk:9.0 AS build
+ARG BUILD_CONFIGURATION=Release
+WORKDIR /src
+COPY ["ServerApp/ServerApp.csproj", "ServerApp/"]
+RUN dotnet restore "ServerApp/ServerApp.csproj"
+COPY . .
+WORKDIR "/src/ServerApp"
+RUN dotnet build "./ServerApp.csproj" -c $BUILD_CONFIGURATION -o /app/build
+
+FROM build AS publish
+ARG BUILD_CONFIGURATION=Release
+RUN dotnet publish "./ServerApp.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "ServerApp.dll"]
