+++ 351004/Kuchko/ServerApp/ServerApp/ServerApp.http
@@ -1,0 +1,6 @@
+@ServerApp_HostAddress = http://localhost:5070
+
+GET {{ServerApp_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
